+++ source/Examples/Resources/xDnsServerSetting/2-xDnsServerSetting_RemoteNode_Config.ps1
@@ -51,7 +51,6 @@ Configuration xDnsServerSetting_RemoteNode_Config
             DnsServer           = 'dns1.company.local'
             ListenAddresses    = '10.0.0.4'
             IsSlave            = $true
-            Forwarders         = @('168.63.129.16', '168.63.129.18')
             RoundRobin         = $true
             LocalNetPriority   = $true
             SecureResponses    = $true