+++ CHANGELOG.md
@@ -5,6 +5,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Removed
+
+- xDnsServerSetting
+  - BREAKING CHANGE: The properties `Forwarders` and `ForwardingTimeout` has
+    been removed ([issue #192](https://github.com/dsccommunity/xDnsServer/issues/192)).
+    Use the resource _xDnsServerForwarder_ to enforce these properties.
+
 ## [2.0.0] - 2021-03-26
 
 ### Deprecated